\name{fourier-package}
\alias{fourier-package}
\alias{fourier}
\docType{package}
\title{\packageTitle{fourier}}
\description{\packageDescription{fourier}}
\section{Package Content}{\packageIndices{fourier}}
\author{\packageAuthor{fourier}}
\section{Maintainer}{\packageMaintainer{fourier}}
\keyword{package}
