% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_amplitude}
\alias{get_amplitude}
\title{Get Amplitude of the Center of Mass of the Wrapped Signal}
\usage{
get_amplitude(x, y)
}
\arguments{
\item{x, y}{Vectors of X and Y coordinates}
}
\value{
amplitude of the relevant frequency used to produce the X and Y
coordinates in the original signal.
}
\description{
For a certain signal at a frequency, the signal is "wrapped" around the
origin at a specific frequency. The magnitude of the center of mass of that
wrapped function is half of the amplitude of that frequency in the transform.
}
