% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_peaks}
\alias{get_peaks}
\title{Get Peaks}
\usage{
get_peaks(data, decline = 0.3, window = 15)
}
\arguments{
\item{data}{vector of numeric data}

\item{decline}{optional, between (0, 1]. Fraction of point P which nearby
points are expected be lower than if P is a true peak (e.g. If the
adjacent points to P are lower than 0.8*P, then we will consider P a peak)}

\item{window}{optional, number of points on either side surrounding a point P
which are expected to be lower than decline*P if P is a true peak.}
}
\value{
Return the indices within data of the local maxima discovered
}
\description{
Implements a local maximum finding algorithm, with certain tolerances for
how sharp the peak is, and how spread apart peaks must be.
}
\examples{
data <- c(1:5, 20, 6:1, rep(5, 5), 6, rep(5,5))
## Should only return 6, since the element, 20, at the 6th index, has at
## least 4 elements on either side which are at least 20\% lower. Even though
## element 18 also is a local maximum, it is not sufficiently higher than its
## neighbors
get_peaks(data, 0.2, 4)
}
