% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dft}
\alias{dft}
\title{Discrete Fourier Transform}
\usage{
dft(data, freq_seq = NULL, sampling_rate = NULL)
}
\arguments{
\item{data}{vector containing a numeric time series in the time-intensity
domain to be transformed to the frequency-amplitude domain}

\item{freq_seq}{vector of}
}
\value{

}
\description{
Discrete Fourier Transform
}
